moved duplicated function into file
